requirements.txt
asgiref==3.7.2
Babel==2.12.1
certifi==2023.5.7
charset-normalizer==3.1.0
django-admin-thumbnails==0.2.8
django-debug-toolbar==4.1.0
django-phonenumber-field==7.1.0
idna==3.4
phonenumbers==8.13.13
Pillow==9.5.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
altgraph==0.17.3
asgiref==3.7.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.1
django-cors-headers==4.3.1
django-debug-toolbar==4.1.0
django-filter==23.3
django-import-export==3.3.3
django-simple-history==3.4.0
djangorestframework==3.14.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
future==0.18.2
inflection==0.5.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
pefile==2022.5.30
pyinstaller==5.6.2
pyinstaller-hooks-contrib==2022.13
python-decouple==3.8
pytz==2023.3.post1
pywin32-ctypes==0.2.0
PyYAML==6.0.1
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
xlrd==2.0.1
xlwt==1.3.0
